+++ torchtune/modules/model_fusion/__init__.py
@@ -4,7 +4,9 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from ._fusion import DeepFusionModel, EarlyFusionModel, FusionEmbedding, FusionLayer
+from ._deep_fusion import DeepFusionModel
+from ._early_fusion import EarlyFusionModel
+from ._fusion_layers import FusionEmbedding, FusionLayer
 from ._fusion_utils import get_fusion_params, register_fusion_module
 
 __all__ = [
